Move photo bound parsing to separate helper class
